+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Implement color channel selector.
 - Implement brush hardness.
 - Implement brush strength.
+- Implement brush density.
 
 ### Changed